chore: Split `PipelineFunctions` into multiple files grouped by Orchestrator responsibility
