+++ platform/src/apis/Platform.Orchestrator/Telemetry/TelemetryService.cs
@@ -1,0 +1,37 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Microsoft.ApplicationInsights;
+
+namespace Platform.Orchestrator.Telemetry;
+
+public interface ITelemetryService
+{
+    void TrackEvent(string eventName, string? id, Dictionary<string, string?>? props = null);
+}
+
+public class TelemetryService(TelemetryClient telemetry) : ITelemetryService
+{
+    public void TrackEvent(string eventName, string? id, Dictionary<string, string?>? props = null)
+    {
+        if (string.IsNullOrWhiteSpace(id))
+        {
+            return;
+        }
+
+        var properties = new Dictionary<string, string>
+        {
+            { "Id", id }
+        };
+
+        if (props != null)
+        {
+            foreach (var prop in props.Where(p => !string.IsNullOrWhiteSpace(p.Value)))
+            {
+                properties[prop.Key] = prop.Value!;
+            }
+        }
+
+        var metrics = new Dictionary<string, double>();
+        telemetry.TrackEvent(eventName, properties, metrics);
+    }
+}
